FRP Overhaul: unified context, modular handlers
Overhauling App so that everything (absolutely everything) is FRP
instead of IORef; this means that not everything will be implemented
fully or correctly the first time, might need some debugging (hence
Test.hs)

Separated out MultiSelect from Widgets.hs and successfully implemented
the right JQuery commands so that it (mostly) works; selections are not
cleared when the button is clicked, but this may yet be changed

Removed App/Core.hs, which served only to provide the now-defunct applet
functionality; having a universal context is simpler and more compatible
with FRP, as well as easier to edit and correct errors.

Added App/Widets/Ranger.hs, which obviates the "<"/">" buttons and
numbered pages; this also makes it possible to use arbitrary enumerable
data types, not just ints, and is more FRP-compatible

Added Ranger to Test.hs; now you can move around through a selection of Shakespeare's
works (very simple, but easily extendable)

Next Steps:
Implementing a CRUD-like mutable dynamic list database for tags,
and an immutable dynamic list database for the card softlinks (might
want to rehash softlinks/liquidlinks for this purpose)

Testing the thing.